Move some svelte components from global folder
