+++ ui/src/components/table/HeadCol.svelte
