+++ ui/src/components/table/TableNav.svelte
