+++ ui/src/pages/admin/Subscriptions.svelte
@@ -4,18 +4,18 @@
   import LL from '../../i18n/i18n-svelte';
   import { appRoutes } from '../../config';
   import { validateUserIsAdmin } from '../../validationUtils';
-  import HeadCol from '../../components/global/table/HeadCol.svelte';
-  import RowCol from '../../components/global/table/RowCol.svelte';
-  import TableRow from '../../components/global/table/TableRow.svelte';
+  import HeadCol from '../../components/table/HeadCol.svelte';
+  import RowCol from '../../components/table/RowCol.svelte';
+  import TableRow from '../../components/table/TableRow.svelte';
   import CheckIcon from '../../components/icons/CheckIcon.svelte';
-  import Table from '../../components/global/table/Table.svelte';
-  import AdminPageLayout from '../../components/global/AdminPageLayout.svelte';
+  import Table from '../../components/table/Table.svelte';
+  import AdminPageLayout from '../../components/AdminPageLayout.svelte';
   import SubscriptionForm from '../../components/subscription/SubscriptionForm.svelte';
   import DeleteConfirmation from '../../components/global/DeleteConfirmation.svelte';
   import HollowButton from '../../components/global/HollowButton.svelte';
-  import TableNav from '../../components/global/table/TableNav.svelte';
-  import TableContainer from '../../components/global/table/TableContainer.svelte';
-  import TableFooter from '../../components/global/table/TableFooter.svelte';
+  import TableNav from '../../components/table/TableNav.svelte';
+  import TableContainer from '../../components/table/TableContainer.svelte';
+  import TableFooter from '../../components/table/TableFooter.svelte';
 
   export let xfetch;
   export let router;
