+++ ui/src/pages/admin/Users.svelte
@@ -5,11 +5,11 @@
   import LL from '../../i18n/i18n-svelte';
   import { AppConfig, appRoutes } from '../../config';
   import { validateUserIsAdmin } from '../../validationUtils';
-  import Table from '../../components/global/table/Table.svelte';
-  import HeadCol from '../../components/global/table/HeadCol.svelte';
-  import AdminPageLayout from '../../components/global/AdminPageLayout.svelte';
-  import TableRow from '../../components/global/table/TableRow.svelte';
-  import RowCol from '../../components/global/table/RowCol.svelte';
+  import Table from '../../components/table/Table.svelte';
+  import HeadCol from '../../components/table/HeadCol.svelte';
+  import AdminPageLayout from '../../components/AdminPageLayout.svelte';
+  import TableRow from '../../components/table/TableRow.svelte';
+  import RowCol from '../../components/table/RowCol.svelte';
   import UserAvatar from '../../components/user/UserAvatar.svelte';
   import CountryFlag from '../../components/user/CountryFlag.svelte';
   import VerifiedIcon from '../../components/icons/VerifiedIcon.svelte';
@@ -17,9 +17,9 @@
   import Modal from '../../components/global/Modal.svelte';
   import DeleteConfirmation from '../../components/global/DeleteConfirmation.svelte';
   import CreateUser from '../../components/user/CreateUser.svelte';
-  import TableContainer from '../../components/global/table/TableContainer.svelte';
-  import TableNav from '../../components/global/table/TableNav.svelte';
-  import TableFooter from '../../components/global/table/TableFooter.svelte';
+  import TableContainer from '../../components/table/TableContainer.svelte';
+  import TableNav from '../../components/table/TableNav.svelte';
+  import TableFooter from '../../components/table/TableFooter.svelte';
 
   export let xfetch;
   export let router;
